+++ app/order/biz/dal/db/ent/schema/order_pay.go
@@ -17,8 +17,8 @@ type OrderPay struct {
 func (OrderPay) Fields() []ent.Field {
 	return []ent.Field{
 		field.Int64("order_id").Comment("订单id").Optional(),
-		field.Float("remission").Comment("减免").Optional(),
-		field.Float("pay").Comment("实际付款").Optional(),
+		field.Int64("remission").Comment("减免").Optional(),
+		field.Int64("pay").Comment("实际付款").Optional(),
 		field.String("note").Comment("备注").Optional(),
 		field.Time("pay_at").Comment("支付时间").Optional(),
 		field.String("pay_way").Comment("支付方式").Optional(),
